Migrate the GKE Usage Metering connector to dscc-gen
