+++ gke-usage-metering/.clasp.json
@@ -1,0 +1,1 @@
+{"scriptId":"1da_9E3RGqzpYqeHSA5wcU6MUDxcEHGDgapiMEbPMLud2e1IghOZAfNSl","rootDir":"src"}
