--- gke-usage-metering/README.md
@@ -1,13 +1,0 @@
-# GKE Usage Metering Community Connector for Data Studio
-
-*This is not an official Google product*
-
-This [Data Studio](https://datastudio.google.com) [Community
-Connector](https://developers.google.com/datastudio/connector) lets users query
-[GKE Usage Metering](https://cloud.google.com/kubernetes-engine/docs/how-to/cluster-usage-metering)
-datasets for visualization.
-
-## Deploy the Community Connector yourself
-
-Use the [deployment guide](../deploy.md) to deploy the Community Connector
-yourself.
