--- gke-usage-metering/jsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "noLib": true
-  }
-}
