--- gke-usage-metering/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": [ "es6" ],
-    "outDir": "build/"
-  },
-  "include": [
-    "data-studio.d.ts",
-    "gkeUsageMetering.ts",
-    "main.ts",
-    "query.ts"
-  ]
-}
